+++ src/fixIcelandicLocale.privates.ts
@@ -1,5 +1,8 @@
 type IntlClassLike = {
-  supportedLocalesOf: (locales: string | Array<string>) => Array<string>;
+  supportedLocalesOf: (
+    locales: string | Array<string>,
+    options?: Record<string, unknown>
+  ) => Array<string>;
 };
 
 const islLocaleRe = /^isl?(?:-|$)/i;
@@ -26,6 +29,34 @@ const mapLocales = (
       return; // no mapping needed. YOLO!
     }
   }
+};
+
+const patchSupportedLocalesOf = <T extends IntlClassLike>(
+  constr: T
+): T['supportedLocalesOf'] => {
+  const BASE_CHAR_CODE = 64; // 'A'.charCodeAt(0) - 1; // used for generating unique suffix for fake locales
+  const sLO: T['supportedLocalesOf'] = (locales, options) => {
+    let localesArr = typeof locales === 'string' ? [locales] : locales;
+    const memoIsl: Array<string> = [];
+    localesArr = localesArr.map((locale) => {
+      if (islLocaleRe.test(locale)) {
+        // Some engines throw a RangeError if the locale is weirdly shaped,
+        // so we must use a short, safe, unique fake locale instead,
+        // and store the actual locale in `memoIsl` for later reinsertion.
+        memoIsl.push(locale);
+        return `da-X${String.fromCharCode(BASE_CHAR_CODE + memoIsl.length)}`;
+      }
+      return locale;
+    });
+    const supportedLocales = constr.supportedLocalesOf(localesArr, options);
+    if (!memoIsl.length) {
+      return supportedLocales;
+    }
+    return supportedLocales.map((locale) =>
+      locale.startsWith('da-X') ? memoIsl.shift()! : locale
+    );
+  };
+  return sLO;
 };
 
 const combineParts = (parts: Array<{ value: string }>) =>
@@ -71,8 +102,7 @@ const PatchedCollator = function Collator(
 };
 
 PatchedCollator.prototype = { constructor: PatchedCollator };
-// Static methods (not patched since "is" is not ACTUALLY supported.)
-PatchedCollator.supportedLocalesOf = _Collator.supportedLocalesOf;
+PatchedCollator.supportedLocalesOf = /*#__PURE__*/ patchSupportedLocalesOf(_Collator);
 PatchedCollator.$original = _Collator;
 
 export const _PatchedCollator = PatchedCollator as unknown as typeof Intl.Collator & {
@@ -152,8 +182,8 @@ const PatchedNumberFormat = function NumberFormat(
 };
 
 PatchedNumberFormat.prototype = { constructor: PatchedNumberFormat };
-// Static methods (not patched since "is" is not ACTUALLY supported.)
-PatchedNumberFormat.supportedLocalesOf = _NumberFormat.supportedLocalesOf;
+PatchedNumberFormat.supportedLocalesOf =
+  /*#__PURE__*/ patchSupportedLocalesOf(_NumberFormat);
 PatchedNumberFormat.$original = _NumberFormat;
 
 export const _PatchedNumberFormat =
@@ -321,8 +351,8 @@ const PatchedDateTimeFormat = function DateTimeFormat(
 };
 
 PatchedDateTimeFormat.prototype = { constructor: PatchedDateTimeFormat };
-// Static methods (not patched since "is" is not ACTUALLY supported.)
-PatchedDateTimeFormat.supportedLocalesOf = _DateTimeFormat.supportedLocalesOf;
+PatchedDateTimeFormat.supportedLocalesOf =
+  /*#__PURE__*/ patchSupportedLocalesOf(_DateTimeFormat);
 PatchedDateTimeFormat.$original = _DateTimeFormat;
 
 export const _PatchedDateTimeFormat =
@@ -461,6 +491,7 @@ if (_PluralRules) {
       return super.selectRange(n, n2);
     }
 
+    static supportedLocalesOf = /*#__PURE__*/ patchSupportedLocalesOf(_PluralRules);
     static $original = _PluralRules;
   };
 }
@@ -506,6 +537,7 @@ if (_ListFormat) {
       return parts;
     }
 
+    static supportedLocalesOf = /*#__PURE__*/ patchSupportedLocalesOf(_ListFormat);
     static $original = _ListFormat;
   };
 }